requirements.txt
openai==1.65.2
python-dotenv==1.0.1
streamlit==1.40.1
pydantic==2.10.6
tenacity==9.0.0
urllib3==2.2.3